fix: event details failed to handle super event  fetching
LIIKUNTA-514
