+++ packages/components/src/hooks/index.ts
@@ -26,3 +26,4 @@ export { default as useAppSportsTranslation } from './useAppSportsTranslation';
 export { default as useCommonCmsConfig } from './useCommonCmsConfig';
 export { default as useClickCapture } from './useClickCapture';
 export { default as usePageScrollRestoration } from './usePageScrollRestoration';
+export { default as useSuperEventLazyLoad } from './useSuperEventLazyLoad';
